Move content item repository to its own file
